+++ frontend/src/app/(site)/[locale]/character/[code]/error.tsx
