+++ frontend/src/app/(site)/[locale]/character/[code]/loading.tsx
