+++ frontend/src/app/(site)/[locale]/patches/[version]/page.tsx
