+++ frontend/src/app/(site)/[locale]/privacy/page.tsx
