+++ frontend/src/app/(site)/[locale]/synergy-detail/error.tsx
