+++ frontend/src/app/(site)/[locale]/synergy-detail/loading.tsx
