+++ frontend/src/app/(site)/[locale]/synergy-detail/page.tsx
