+++ frontend/src/app/(site)/[locale]/synergy/page.tsx
