+++ frontend/src/app/(site)/[locale]/terms/page.tsx
