+++ frontend/src/app/(site)/[locale]/updates/page.tsx
