+++ frontend/src/lib/geistFont.ts
@@ -1,0 +1,6 @@
+import { Geist } from "next/font/google";
+
+export const geistSans = Geist({
+  variable: "--font-geist-sans",
+  subsets: ["latin"],
+});
